--- apps/upskii/src/app/[locale]/(dashboard)/[wsId]/(admin)/roles/client-page.tsx
@@ -1,40 +1,0 @@
-'use client';
-
-import { addWhitelistEmail } from './actions';
-import WhitelistEmailForm from './form';
-import { useToast } from '@tuturuuu/ui/hooks/use-toast';
-import { useTranslations } from 'next-intl';
-import { useRouter } from 'next/navigation';
-
-interface Props {
-  wsId: string;
-  onFinish?: () => void;
-}
-
-export default function WhitelistEmailClient({ wsId, onFinish }: Props) {
-  const t = useTranslations();
-  const router = useRouter();
-
-  const { toast } = useToast();
-
-  const handleSubmit = async (values: { email: string }) => {
-    try {
-      await addWhitelistEmail(values.email, true);
-      toast({
-        title: t('common.success'),
-        description: t('common.email_added'),
-      });
-      onFinish?.();
-      router.refresh();
-      // eslint-disable-next-line no-unused-vars
-    } catch (error) {
-      toast({
-        title: t('common.error'),
-        description: t('common.error_adding_email'),
-        variant: 'destructive',
-      });
-    }
-  };
-
-  return <WhitelistEmailForm wsId={wsId} onSubmit={handleSubmit} />;
-}
