--- apps/upskii/src/app/[locale]/(dashboard)/[wsId]/(admin)/roles/form.tsx
@@ -1,72 +1,0 @@
-import { Button } from '@tuturuuu/ui/button';
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@tuturuuu/ui/form';
-import { useForm } from '@tuturuuu/ui/hooks/use-form';
-import { Input } from '@tuturuuu/ui/input';
-import { zodResolver } from '@tuturuuu/ui/resolvers';
-import { useTranslations } from 'next-intl';
-import * as z from 'zod';
-
-interface Props {
-  wsId: string;
-  // eslint-disable-next-line no-unused-vars
-  onSubmit: (values: z.infer<typeof FormSchema>) => void;
-}
-
-const FormSchema = z.object({
-  email: z.string().email(),
-});
-
-export const WhitelistEmailFormSchema = FormSchema;
-
-export default function WhitelistEmailForm({ onSubmit }: Props) {
-  const t = useTranslations();
-
-  const form = useForm({
-    resolver: zodResolver(FormSchema),
-    defaultValues: {
-      email: '',
-    },
-  });
-
-  const isDirty = form.formState.isDirty;
-  const isValid = form.formState.isValid;
-  const isSubmitting = form.formState.isSubmitting;
-
-  const disabled = !isDirty || !isValid || isSubmitting;
-
-  return (
-    <Form {...form}>
-      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-3">
-        <FormField
-          control={form.control}
-          name="email"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Email</FormLabel>
-              <FormControl>
-                <Input
-                  type="email"
-                  placeholder="example@email.com"
-                  autoComplete="off"
-                  {...field}
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <Button type="submit" className="w-full" disabled={disabled}>
-          {t('ai-whitelist.add_email')}
-        </Button>
-      </form>
-    </Form>
-  );
-}
