+++ apps/upskii/src/app/[locale]/(dashboard)/[wsId]/(admin)/roles/form/role-permissions.tsx
