--- apps/upskii/src/app/[locale]/(dashboard)/[wsId]/(admin)/roles/layout.tsx
@@ -1,36 +1,0 @@
-import {
-  createAdminClient,
-  createClient,
-} from '@tuturuuu/supabase/next/server';
-import { redirect } from 'next/navigation';
-
-export default async function RootLayout({
-  children,
-  params,
-}: {
-  children: React.ReactNode;
-  params: Promise<{
-    wsId: string;
-  }>;
-}) {
-  const { wsId } = await params;
-  const sbAdmin = await createAdminClient();
-  const supabase = await createClient();
-
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user?.email) redirect('/login');
-
-  const { data: whitelisted } = await sbAdmin
-    .from('nova_roles')
-    .select('enabled,  allow_role_management')
-    .eq('email', user.email)
-    .maybeSingle();
-
-  if (!whitelisted?.enabled || !whitelisted?.allow_role_management)
-    redirect(`/${wsId}/home`);
-
-  return children;
-}
